+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Updated/Enhanced Rules and Bug Fixes
  - #408: Fix a typo in **ClosureAsLastMethodParameter** rule's violation message. Fixes #402. (Marcin Erdmann)
  - #410: **IndentationRule**: Fix NPE on GStrings containing an if-statement. (debanne)
  - #398: **ClassStartsWithBlankLine** and **ClassEndsWithBlankLine** rules: Ignore Script classes.
+ - #399: **AssignCollectionUnique** rule: Fix false positive for `unique(false)`, `unique(false, Comparator)` and `unique(false, Closure)`.
 
 Tests and Documentation
  - #411: Consistent indentation on IndentationRuleTest.groovy. (debanne)